model/requirements.txt
flask==2.0.1
tensorflow==2.9.1
numpy==1.23.5
pillow==9.3.0
flask-cors==3.0.10
gunicorn==20.1.0
scikit-learn==1.0.2
matplotlib==3.5.3
pandas==1.5.3 